# added changes to do less i/o to Redis