--- a/middleware.go
+++ b/middleware.go
@@ -269,10 +269,17 @@ func (t BaseMiddleware) ApplyPolicies(key string, session *user.SessionState) er
 			tags[tag] = true
 		}
 	}
-	session.Tags = make([]string, 0, len(tags))
-	for tag := range tags {
-		session.Tags = append(session.Tags, tag)
+
+	// set tags
+	if len(tags) > 0 {
+		session.Tags = make([]string, 0, len(tags))
+		for tag := range tags {
+			session.Tags = append(session.Tags, tag)
+		}
+	} else {
+		session.Tags = nil
 	}
+
 	session.AccessRights = rights
 	// Update the session in the session manager in case it gets called again
 	return t.Spec.SessionManager.UpdateSession(key, session, session.Lifetime(t.Spec.SessionLifetime), false)