--- a/mw_auth_key_test.go
+++ b/mw_auth_key_test.go
@@ -10,10 +10,12 @@ import (
 
 	"github.com/justinas/alice"
 
+	"github.com/lonelycode/go-uuid/uuid"
+
 	"github.com/TykTechnologies/tyk/user"
 )
 
-func createAuthKeyAuthSession() *user.SessionState {
+func createAuthKeyAuthSession(isBench bool) *user.SessionState {
 	session := new(user.SessionState)
 	// essentially non-throttled
 	session.Rate = 100.0
@@ -22,8 +24,13 @@ func createAuthKeyAuthSession() *user.SessionState {
 	session.Per = 1.0
 	session.QuotaRenewalRate = 300 // 5 minutes
 	session.QuotaRenews = time.Now().Unix()
-	session.QuotaRemaining = 10
-	session.QuotaMax = 10
+	if isBench {
+		session.QuotaRemaining = 100000000
+		session.QuotaMax = 100000000
+	} else {
+		session.QuotaRemaining = 10
+		session.QuotaMax = 10
+	}
 	session.AccessRights = map[string]user.AccessDefinition{"31": {APIName: "Tyk Auth Key Test", APIID: "31", Versions: []string{"default"}}}
 	return session
 }
@@ -45,17 +52,22 @@ func getAuthKeyChain(spec *APISpec) http.Handler {
 	return chain
 }
 
-func testPrepareAuthKeySession(tb testing.TB, apiDef string) (string, *APISpec) {
+func testPrepareAuthKeySession(tb testing.TB, apiDef string, isBench bool) (string, *APISpec) {
 	spec := createSpecTest(tb, apiDef)
-	session := createAuthKeyAuthSession()
-	customToken := "54321111"
+	session := createAuthKeyAuthSession(isBench)
+	customToken := ""
+	if isBench {
+		customToken = uuid.New()
+	} else {
+		customToken = "54321111"
+	}
 	// AuthKey sessions are stored by {token}
 	spec.SessionManager.UpdateSession(customToken, session, 60, false)
 	return customToken, spec
 }
 
 func TestBearerTokenAuthKeySession(t *testing.T) {
-	customToken, spec := testPrepareAuthKeySession(t, authKeyDef)
+	customToken, spec := testPrepareAuthKeySession(t, authKeyDef, false)
 
 	recorder := httptest.NewRecorder()
 	req := testReq(t, "GET", "/auth_key_test/", nil)
@@ -74,7 +86,7 @@ func TestBearerTokenAuthKeySession(t *testing.T) {
 func BenchmarkBearerTokenAuthKeySession(b *testing.B) {
 	b.ReportAllocs()
 
-	customToken, spec := testPrepareAuthKeySession(b, authKeyDef)
+	customToken, spec := testPrepareAuthKeySession(b, authKeyDef, true)
 
 	recorder := httptest.NewRecorder()
 	req := testReq(b, "GET", "/auth_key_test/", nil)
@@ -109,7 +121,7 @@ const authKeyDef = `{
 }`
 
 func TestMultiAuthBackwardsCompatibleSession(t *testing.T) {
-	customToken, spec := testPrepareAuthKeySession(t, multiAuthBackwardsCompatible)
+	customToken, spec := testPrepareAuthKeySession(t, multiAuthBackwardsCompatible, false)
 
 	recorder := httptest.NewRecorder()
 	req := testReq(t, "GET", fmt.Sprintf("/auth_key_test/?token=%s", customToken), nil)
@@ -126,7 +138,7 @@ func TestMultiAuthBackwardsCompatibleSession(t *testing.T) {
 func BenchmarkMultiAuthBackwardsCompatibleSession(b *testing.B) {
 	b.ReportAllocs()
 
-	customToken, spec := testPrepareAuthKeySession(b, multiAuthBackwardsCompatible)
+	customToken, spec := testPrepareAuthKeySession(b, multiAuthBackwardsCompatible, true)
 
 	recorder := httptest.NewRecorder()
 	req := testReq(b, "GET", fmt.Sprintf("/auth_key_test/?token=%s", customToken), nil)
@@ -163,7 +175,7 @@ const multiAuthBackwardsCompatible = `{
 
 func TestMultiAuthSession(t *testing.T) {
 	spec := createSpecTest(t, multiAuthDef)
-	session := createAuthKeyAuthSession()
+	session := createAuthKeyAuthSession(false)
 	customToken := "54321111"
 	// AuthKey sessions are stored by {token}
 	spec.SessionManager.UpdateSession(customToken, session, 60, false)