@matiasinsaurralde 2018-05-19
Seems like `BenchmarkMultiSession_BA_Standard_OK` is impacted by the new murmur calls?