@dencoded 2018-05-19
@buger @matiasinsaurralde  I've been looking into this problem as well. I don't think it is because now we calculate murmur for real. If I run it standalone - benchcmp shows pretty descent results:
```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkMultiSession_BA_Standard_OK-8     5586111       4095055       -26.69%

benchmark                                  old allocs     new allocs     delta
BenchmarkMultiSession_BA_Standard_OK-8     428            284            -33.64%

benchmark                                  old bytes     new bytes     delta
BenchmarkMultiSession_BA_Standard_OK-8     83567         27588         -66.99%
```

I think that problem connected with fact that now we don't touch the session if it is not changed and the way we create session in our tests - I feel like tests influence each other. I will push a fix shortly.