@dencoded 2018-05-20
confirmed that with these changes our rate/quota limiter blocks `BenchmarkMultiSession_BA_Standard_OK` from running fast, it returns error `Error: Quota exceeded errCode: 403` for all (except 1st 10 calls) and gives most of work to `ErrorHandler` which has lots of ops inside.

if we look at:
```
func createMultiAuthKeyAuthSession() *user.SessionState {
	session := new(user.SessionState)
	session.Rate = 100.0
	session.Allowance = session.Rate
	session.LastCheck = time.Now().Unix()
	session.Per = 1.0
	session.QuotaRenewalRate = 300 // 5 minutes
	session.QuotaRenews = time.Now().Unix()
	session.QuotaRemaining = 900
	session.QuotaMax = 10
	session.AccessRights = map[string]user.AccessDefinition{"55": {APIName: "Tyk Multi Key Test", APIID: "55", Versions: []string{"default"}}}
	return session
}
```

we see `session.QuotaMax = 10` and `session.QuotaRenewalRate = 300 // 5 minutes` - it stops `BenchmarkMultiSession_BA_Standard_OK` from normal run for sure (just ten request for 5 minutes while benchmark is doing several Ks per run). however, on master this benchmark runs normally.

I think we need to understand why this benchmark works OK on master - looks like it just ignores quota set for session.