@buger 2018-05-22
Also, It will be fair for benchmarks if we create hundreds of keys, which limited like this, and benchmark will randomly choose between them, instead of using 1 session.