@dencoded 2018-05-23
@buger  I've added a couple of changes - `BenchmarkMultiSession_BA_Standard_OK` (and some other) look better now:
```
benchmark                                                   old ns/op     new ns/op     delta
BenchmarkURLReplacer-8                                      19036         18703         -1.75%
BenchmarkTagHeaders-8                                       975           962           -1.33%
BenchmarkDefaultVersion-8                                   1518035       1101265       -27.45%
BenchmarkGetVersionFromRequest/Header_location-8            534965        498511        -6.81%
BenchmarkGetVersionFromRequest/URL_param_location-8         501352        490345        -2.20%
BenchmarkGetVersionFromRequest/URL_location-8               501976        472858        -5.80%
BenchmarkApiReload-8                                        3811360       3497692       -8.23%
BenchmarkValueExtractor/HeaderSource-8                      125998        126157        +0.13%
BenchmarkValueExtractor/FormSource-8                        138120        139929        +1.31%
BenchmarkRegexExtractor/HeaderSource-8                      132173        158059        +19.58%
BenchmarkRegexExtractor/BodySource-8                        134681        142991        +6.17%
BenchmarkRegexExtractor/FormSource-8                        145787        160229        +9.91%
BenchmarkXPathExtractor/HeaderSource-8                      151543        160926        +6.19%
BenchmarkXPathExtractor/BodySource-8                        153866        163758        +6.43%
BenchmarkXPathExtractor/FormSource-8                        164672        160720        -2.40%
BenchmarkInitGenericEventHandlers-8                         808           742           -8.17%
BenchmarkMultiSession_BA_Standard_OK-8                      1591845       1539256       -3.30%
BenchmarkBearerTokenAuthKeySession-8                        1828327       1430046       -21.78%
BenchmarkMultiAuthBackwardsCompatibleSession-8              1852510       1940279       +4.74%
BenchmarkBasicAuth-8                                        164595920     162841104     -1.07%
BenchmarkContextVarsMiddleware-8                            1393747       1418655       +1.79%
BenchmarkMiddlewareContextVars_ProcessRequest_cookies-8     149227        139889        -6.26%
BenchmarkHMACAuthSessionPass-8                              1601564       1340059       -16.33%
BenchmarkHMACAuthSessionPassWithHeaderField-8               1566401       1364649       -12.88%
BenchmarkIPBlacklistMiddleware-8                            28209         28077         -0.47%
BenchmarkIPMiddlewarePass-8                                 18202         18189         -0.07%
BenchmarkJWTSessionHMAC-8                                   681418        393608        -42.24%
BenchmarkJWTSessionRSA-8                                    814621        515433        -36.73%
BenchmarkJWTSessionRSABearer-8                              807150        502191        -37.78%
BenchmarkJWTSessionRSAWithRawSourceOnWithClientID-8         747932        523018        -30.07%
BenchmarkJWTSessionRSAWithRawSource-8                       740298        505002        -31.78%
BenchmarkJWTSessionRSAWithJWK-8                             726428        557022        -23.32%
BenchmarkJWTSessionRSAWithEncodedJWK-8                      731198        514025        -29.70%
BenchmarkProcessRequestLiveQuotaLimit-8                     636951        664417        +4.31%
BenchmarkProcessRequestOffThreadQuotaLimit-8                473423        463918        -2.01%
BenchmarkProcessRequestLiveRedisRollingLimiter-8            985372        914559        -7.19%
BenchmarkProcessRequestOffThreadRedisRollingLimiter-8       780984        756432        -3.14%
BenchmarkStripAuth_stripFromHeaders-8                       9003          9260          +2.85%
BenchmarkStripAuth_stripFromParams-8                        41911         39982         -4.60%
BenchmarkTransformNonAscii-8                                16314         15879         -2.67%
BenchmarkTransformJSONMarshal-8                             17156         16961         -1.14%
BenchmarkRewriter-8                                         58368         56902         -2.51%
BenchmarkValidateJSONSchema-8                               1866378       1811239       -2.95%
BenchmarkVersioning-8                                       2383908       2082370       -12.65%
BenchmarkVirtualEndpoint-8                                  369881        356590        -3.59%
BenchmarkApplyPolicies-8                                    27422         26838         -2.13%
BenchmarkResponseHeaderInjection-8                          2160559       2123354       -1.72%
BenchmarkCopyRequestResponse-8                              15688         15603         -0.54%

benchmark                                                   old allocs     new allocs     delta
BenchmarkURLReplacer-8                                      44             44             +0.00%
BenchmarkTagHeaders-8                                       15             15             +0.00%
BenchmarkDefaultVersion-8                                   1724           1637           -5.05%
BenchmarkGetVersionFromRequest/Header_location-8            676            655            -3.11%
BenchmarkGetVersionFromRequest/URL_param_location-8         639            639            +0.00%
BenchmarkGetVersionFromRequest/URL_location-8               628            628            +0.00%
BenchmarkApiReload-8                                        23907          23907          +0.00%
BenchmarkValueExtractor/HeaderSource-8                      50             50             +0.00%
BenchmarkValueExtractor/FormSource-8                        77             77             +0.00%
BenchmarkRegexExtractor/HeaderSource-8                      52             52             +0.00%
BenchmarkRegexExtractor/BodySource-8                        62             62             +0.00%
BenchmarkRegexExtractor/FormSource-8                        79             79             +0.00%
BenchmarkXPathExtractor/HeaderSource-8                      108            108            +0.00%
BenchmarkXPathExtractor/BodySource-8                        116            116            +0.00%
BenchmarkXPathExtractor/FormSource-8                        140            140            +0.00%
BenchmarkMultiSession_BA_Standard_OK-8                      464            498            +7.33%
BenchmarkBearerTokenAuthKeySession-8                        387            451            +16.54%
BenchmarkMultiAuthBackwardsCompatibleSession-8              391            446            +14.07%
BenchmarkBasicAuth-8                                        20134          20077          -0.28%
BenchmarkContextVarsMiddleware-8                            1950           1950           +0.00%
BenchmarkMiddlewareContextVars_ProcessRequest_cookies-8     28             28             +0.00%
BenchmarkHMACAuthSessionPass-8                              390            294            -24.62%
BenchmarkHMACAuthSessionPassWithHeaderField-8               415            321            -22.65%
BenchmarkIPBlacklistMiddleware-8                            152            152            +0.00%
BenchmarkIPMiddlewarePass-8                                 95             95             +0.00%
BenchmarkJWTSessionHMAC-8                                   677            571            -15.66%
BenchmarkJWTSessionRSA-8                                    753            703            -6.64%
BenchmarkJWTSessionRSABearer-8                              755            705            -6.62%
BenchmarkJWTSessionRSAWithRawSourceOnWithClientID-8         778            724            -6.94%
BenchmarkJWTSessionRSAWithRawSource-8                       762            713            -6.43%
BenchmarkJWTSessionRSAWithJWK-8                             766            717            -6.40%
BenchmarkJWTSessionRSAWithEncodedJWK-8                      770            691            -10.26%
BenchmarkProcessRequestLiveQuotaLimit-8                     543            556            +2.39%
BenchmarkProcessRequestOffThreadQuotaLimit-8                514            527            +2.53%
BenchmarkProcessRequestLiveRedisRollingLimiter-8            2038           2094           +2.75%
BenchmarkProcessRequestOffThreadRedisRollingLimiter-8       2265           2282           +0.75%
BenchmarkStripAuth_stripFromHeaders-8                       66             66             +0.00%
BenchmarkStripAuth_stripFromParams-8                        220            220            +0.00%
BenchmarkTransformNonAscii-8                                98             98             +0.00%
BenchmarkTransformJSONMarshal-8                             101            101            +0.00%
BenchmarkRewriter-8                                         269            269            +0.00%
BenchmarkValidateJSONSchema-8                               2529           2528           -0.04%
BenchmarkVersioning-8                                       2927           2860           -2.29%
BenchmarkVirtualEndpoint-8                                  372            372            +0.00%
BenchmarkApplyPolicies-8                                    102            102            +0.00%
BenchmarkResponseHeaderInjection-8                          2508           2507           -0.04%
BenchmarkCopyRequestResponse-8                              76             76             +0.00%

benchmark                                                   old bytes     new bytes     delta
BenchmarkURLReplacer-8                                      849           849           +0.00%
BenchmarkTagHeaders-8                                       384           384           +0.00%
BenchmarkDefaultVersion-8                                   247505        221578        -10.48%
BenchmarkGetVersionFromRequest/Header_location-8            97335         96386         -0.97%
BenchmarkGetVersionFromRequest/URL_param_location-8         96091         96091         +0.00%
BenchmarkGetVersionFromRequest/URL_location-8               93366         93333         -0.04%
BenchmarkApiReload-8                                        1789676       1789639       -0.00%
BenchmarkValueExtractor/HeaderSource-8                      4569          4569          +0.00%
BenchmarkValueExtractor/FormSource-8                        10666         10666         +0.00%
BenchmarkRegexExtractor/HeaderSource-8                      4765          4767          +0.04%
BenchmarkRegexExtractor/BodySource-8                        9213          9214          +0.01%
BenchmarkRegexExtractor/FormSource-8                        10862         10863         +0.01%
BenchmarkXPathExtractor/HeaderSource-8                      9375          9375          +0.00%
BenchmarkXPathExtractor/BodySource-8                        13794         13794         +0.00%
BenchmarkXPathExtractor/FormSource-8                        15986         15987         +0.01%
BenchmarkMultiSession_BA_Standard_OK-8                      90385         83059         -8.11%
BenchmarkBearerTokenAuthKeySession-8                        79254         78572         -0.86%
BenchmarkMultiAuthBackwardsCompatibleSession-8              80415         79938         -0.59%
BenchmarkBasicAuth-8                                        1080680       1062460       -1.69%
BenchmarkContextVarsMiddleware-8                            304276        304263        -0.00%
BenchmarkMiddlewareContextVars_ProcessRequest_cookies-8     2387          2386          -0.04%
BenchmarkHMACAuthSessionPass-8                              79461         47584         -40.12%
BenchmarkHMACAuthSessionPassWithHeaderField-8               80786         50704         -37.24%
BenchmarkIPBlacklistMiddleware-8                            32338         32345         +0.02%
BenchmarkIPMiddlewarePass-8                                 26734         26734         +0.00%
BenchmarkJWTSessionHMAC-8                                   105775        85673         -19.00%
BenchmarkJWTSessionRSA-8                                    135419        109102        -19.43%
BenchmarkJWTSessionRSABearer-8                              136594        110307        -19.24%
BenchmarkJWTSessionRSAWithRawSourceOnWithClientID-8         127023        110085        -13.33%
BenchmarkJWTSessionRSAWithRawSource-8                       121301        107421        -11.44%
BenchmarkJWTSessionRSAWithJWK-8                             121338        107486        -11.42%
BenchmarkJWTSessionRSAWithEncodedJWK-8                      121509        106313        -12.51%
BenchmarkProcessRequestLiveQuotaLimit-8                     87202         85057         -2.46%
BenchmarkProcessRequestOffThreadQuotaLimit-8                85880         83687         -2.55%
BenchmarkProcessRequestLiveRedisRollingLimiter-8            147363        143126        -2.88%
BenchmarkProcessRequestOffThreadRedisRollingLimiter-8       156260        150689        -3.57%
BenchmarkStripAuth_stripFromHeaders-8                       2800          2800          +0.00%
BenchmarkStripAuth_stripFromParams-8                        14161         14161         +0.00%
BenchmarkTransformNonAscii-8                                11060         11060         +0.00%
BenchmarkTransformJSONMarshal-8                             11508         11508         +0.00%
BenchmarkRewriter-8                                         49940         49940         +0.00%
BenchmarkValidateJSONSchema-8                               287817        287775        -0.01%
BenchmarkVersioning-8                                       312095        292117        -6.40%
BenchmarkVirtualEndpoint-8                                  70094         70096         +0.00%
BenchmarkApplyPolicies-8                                    12585         12585         +0.00%
BenchmarkResponseHeaderInjection-8                          400072        400007        -0.02%
BenchmarkCopyRequestResponse-8                              35584         35584         +0.00%
```

if it looks OK - lets merge it and use as a baseline for further performance improvements